requirements.txt
discord.pyaiohttpaiosqlitepython-dotenvcoloramacustomtkinterpackagingpillow